requirements.txt
certifi==2018.8.24
enviroplus==0.0.6
influxdb_client==1.36.0
pimoroni_bme280==0.1.1
pms5003==0.0.5
smbus==1.1.post2
